Seperate VGA code Directory and strings in utils directory
