+++ src/utils/strings/strings.c
@@ -1,0 +1,14 @@
+#include "strings.h"
+
+/*
+returns length of the string
+*/
+size_t strlen(char* str)
+{
+    int count=0;
+    while(*str != '\0'){
+        count++;
+        str++;
+    }
+    return count;
+}
