+++ src/utils/strings/strings.h
@@ -1,0 +1,13 @@
+#ifndef __STRINGS_H__
+#define __STRINGS_H__
+//include
+#include <stddef.h>
+#include <stdint.h>
+
+//Function Prototype
+/*
+Return String length
+*/
+size_t strlen(char* str);
+#endif
+
